+++ modelserviceconn.go
@@ -1,27 +1,16 @@
 package tensorflow_service_apis
 
 import (
-	"context"
-	"time"
-
 	tfserv "github.com/Golang-Tools/tensorflow_service_apis/tensorflow_serving/apis"
 	grpc "google.golang.org/grpc"
+	"google.golang.org/grpc/connectivity"
 )
 
-//ModelServiceConn 客户端类
+//ModelServiceConn ModelServiceClient客户端的连接类
 type ModelServiceConn struct {
 	tfserv.ModelServiceClient
 	conn *grpc.ClientConn
 	sdk  *SDK
-}
-
-//ModelServiceConn 建立一个新的连接
-func (c *SDK) NewModelServiceConn() (*ModelServiceConn, error) {
-	conn, err := newModelServiceConn(c, c.addr, c.opts...)
-	if err != nil {
-		return nil, err
-	}
-	return conn, nil
 }
 
 func newModelServiceConn(sdk *SDK, addr string, opts ...grpc.DialOption) (*ModelServiceConn, error) {
@@ -36,17 +25,40 @@ func newModelServiceConn(sdk *SDK, addr string, opts ...grpc.DialOption) (*Model
 	return c, nil
 }
 
+//NewModelServiceConn 建立一个新的ModelServiceClient客户端的连接类
+func (c *SDK) NewModelServiceConn() (*ModelServiceConn, error) {
+	conn, err := newModelServiceConn(c, c.addr, c.opts...)
+	if err != nil {
+		return nil, err
+	}
+	c.getConnLock.Lock()
+	defer c.getConnLock.Unlock()
+	c.modelServiceConn = conn
+	return conn, nil
+}
+
+func (c *SDK) getModelServiceConn() *ModelServiceConn {
+	c.getConnLock.RLock()
+	defer c.getConnLock.RUnlock()
+	if c.modelServiceConn != nil {
+		if c.modelServiceConn.conn.GetState() == connectivity.Shutdown {
+			return nil
+		}
+		return c.modelServiceConn
+	}
+	return nil
+}
+
+//GetModelServiceConn 获取ModelServiceClient客户端连接
+func (c *SDK) GetModelServiceConn() (*ModelServiceConn, error) {
+	conn := c.getModelServiceConn()
+	if conn != nil {
+		return conn, nil
+	}
+	return c.NewModelServiceConn()
+}
+
 //Close 断开连接
 func (c *ModelServiceConn) Close() error {
 	return c.conn.Close()
-}
-
-func (c *ModelServiceConn) NewCtx() (ctx context.Context, cancel context.CancelFunc) {
-	if c.sdk.SDKConfig.Query_Timeout > 0 {
-		timeout := time.Duration(c.sdk.SDKConfig.Query_Timeout) * time.Millisecond
-		ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	} else {
-		ctx, cancel = context.WithCancel(context.Background())
-	}
-	return
 }