+++ predictionserviceconn.go
@@ -1,27 +1,16 @@
 package tensorflow_service_apis
 
 import (
-	"context"
-	"time"
-
 	tfserv "github.com/Golang-Tools/tensorflow_service_apis/tensorflow_serving/apis"
 	grpc "google.golang.org/grpc"
+	"google.golang.org/grpc/connectivity"
 )
 
-//PredictionServiceConn 客户端类
+//PredictionServiceConn PredictionServiceClient客户端的连接类
 type PredictionServiceConn struct {
 	tfserv.PredictionServiceClient
 	conn *grpc.ClientConn
 	sdk  *SDK
-}
-
-//ModelServiceConn 建立一个新的连接
-func (c *SDK) NewPredictionServiceConn() (*PredictionServiceConn, error) {
-	conn, err := newPredictionServiceConn(c, c.addr, c.opts...)
-	if err != nil {
-		return nil, err
-	}
-	return conn, nil
 }
 
 func newPredictionServiceConn(sdk *SDK, addr string, opts ...grpc.DialOption) (*PredictionServiceConn, error) {
@@ -36,17 +25,40 @@ func newPredictionServiceConn(sdk *SDK, addr string, opts ...grpc.DialOption) (*
 	return c, nil
 }
 
+//NewPredictionServiceConn 建立一个新的连接
+func (c *SDK) NewPredictionServiceConn() (*PredictionServiceConn, error) {
+	conn, err := newPredictionServiceConn(c, c.addr, c.opts...)
+	if err != nil {
+		return nil, err
+	}
+	c.getConnLock.Lock()
+	defer c.getConnLock.Unlock()
+	c.predictionServiceConn = conn
+	return conn, nil
+}
+
+func (c *SDK) getPredictionServiceConn() *PredictionServiceConn {
+	c.getConnLock.RLock()
+	defer c.getConnLock.RUnlock()
+	if c.predictionServiceConn != nil {
+		if c.predictionServiceConn.conn.GetState() == connectivity.Shutdown {
+			return nil
+		}
+		return c.predictionServiceConn
+	}
+	return nil
+}
+
+//GetPredictionServiceConn 获取PredictionServiceClient客户端连接
+func (c *SDK) GetPredictionServiceConn() (*PredictionServiceConn, error) {
+	conn := c.getPredictionServiceConn()
+	if conn != nil {
+		return conn, nil
+	}
+	return c.NewPredictionServiceConn()
+}
+
 //Close 断开连接
 func (c *PredictionServiceConn) Close() error {
 	return c.conn.Close()
-}
-
-func (c *PredictionServiceConn) NewCtx() (ctx context.Context, cancel context.CancelFunc) {
-	if c.sdk.SDKConfig.Query_Timeout > 0 {
-		timeout := time.Duration(c.sdk.SDKConfig.Query_Timeout) * time.Millisecond
-		ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	} else {
-		ctx, cancel = context.WithCancel(context.Background())
-	}
-	return
 }